Add: reset tmp dir setting.

+++ tests/test_create_github_movie.py
@@ -9,7 +9,7 @@ import glob
 # rest image and movie folder.
 for folder in glob.glob("image/*"): shutil.rmtree(folder)
 for file in glob.glob("movie/*.mp4"): os.remove(file)
-shutil.rmtree('tmp')
+if not Path('tmp').exists(): shutil.rmtree('tmp')
 
 
 class GithubMovieMaker:
@@ -26,16 +26,27 @@ class GithubMovieMaker:
         download_url = f'{url}/archive/master.zip'
         project_name = url.split('/')[-1]
         zip_path = Path(f"tmp/{project_name}.zip")
-        os.mkdir(zip_path.parent)
+        os.makedirs(zip_path.parent, exist_ok=True)
         FileHandler.download_file(download_url, zip_path)
         FileHandler.unzip_file(zip_path, zip_path.parent)
-        folder_path = zip_path.parent.glob(f"{project_name}-*").__next__()
+        folder_path = self.get_unzip_folder_path(zip_path)
         return folder_path
 
+    @staticmethod
+    def get_unzip_folder_path(zip_path):
+        """
+        Get unzip folder path
+        :param zip_path:
+        :return: unzip folder path
+        """
+        return zip_path.parent.glob(f"{zip_path.stem}-*").__next__()
 
-@pytest.mark.parametrize(('url'), [
-    ("https://github.com/noricha-vr/screen_capture"),
-])
-def test_extract_zip_file(url):
-    folder_path = GithubMovieMaker().download(url)
-    assert folder_path.exists() is True
+
+class TestGithubMovieMaker:
+
+    @pytest.mark.parametrize(('url'), [
+        ("https://github.com/noricha-vr/screen_capture"),
+    ])
+    def test_extract_zip_file(self, url):
+        folder_path = GithubMovieMaker().download(url)
+        assert folder_path.exists() is True
